refactor(lsp): extract var markdown helpers into shared module
Move format_single, format_multi_def, resolve_to_raw_value,
multi_def_separator and the swatch_markdown helper out of hover.rs
into a new var_markdown module so completion no longer reaches across
into the hover module for them.
