+++ crates/css-var-kit/src/commands/lsp/hover.rs
@@ -1,15 +1,15 @@
 use std::error::Error;
 use std::path::Path;
 
-use base64::Engine as _;
-use base64::engine::general_purpose::STANDARD as BASE64;
 use lightningcss::properties::custom::{TokenList, TokenOrValue, Variable};
 use lsp_server::{Message, Request, Response};
 use lsp_types::{Hover, HoverContents, HoverParams, MarkupContent, MarkupKind, Position, Range};
 
 use super::Server;
 use super::definition::extract_variable_at_cursor;
-use crate::color_value::parse_to_rgba;
+use super::var_markdown::{
+    format_multi_def, format_single, multi_def_separator, resolve_to_raw_value,
+};
 use crate::owned_types::OwnedPropId;
 use crate::parser::Property;
 use crate::searcher::PropMapFor;
@@ -48,13 +48,6 @@ impl Server<'_> {
         let separator = multi_def_separator(self.client_name.as_deref());
 
         compute_hover(source, &pos, &rel_path, &usages, &var_defs, separator)
-    }
-}
-
-pub(super) fn multi_def_separator(client_name: Option<&str>) -> &'static str {
-    match client_name {
-        Some(name) if name.eq_ignore_ascii_case("helix") => "  \n",
-        _ => "\n\n",
     }
 }
 
@@ -137,68 +130,6 @@ fn format_var_hover(
             Some(format_single(&resolved, true))
         }
     }
-}
-
-pub(super) fn format_single(resolved: &str, include_swatch: bool) -> String {
-    match parse_to_rgba(resolved).filter(|_| include_swatch) {
-        Some(color) => format!("{} `{resolved}`", swatch_markdown(&color)),
-        None => format!("`{resolved}`"),
-    }
-}
-
-pub(super) fn format_multi_def(
-    props: &[&Property],
-    var_defs: &PropMapFor<'_, VariableDefinitions>,
-    separator: &str,
-    include_swatch: bool,
-) -> Option<String> {
-    let lines: Vec<String> = props
-        .iter()
-        .filter_map(|prop| {
-            let resolved = resolve_to_raw_value(prop, var_defs, 0)?;
-            let location = format!("{}:{}", prop.file_path.display(), prop.ident.line + 1);
-            Some(match parse_to_rgba(&resolved).filter(|_| include_swatch) {
-                Some(color) => format!("{} `{resolved}` — {location}", swatch_markdown(&color)),
-                None => format!("`{resolved}` — {location}"),
-            })
-        })
-        .collect();
-
-    (!lines.is_empty()).then(|| lines.join(separator))
-}
-
-pub(super) fn resolve_to_raw_value(
-    prop: &Property,
-    var_defs: &PropMapFor<'_, VariableDefinitions>,
-    depth: usize,
-) -> Option<String> {
-    if depth > 32 {
-        return None;
-    }
-    let token_list = prop.token_list().inner();
-    if let [TokenOrValue::Var(var)] = token_list.0.as_slice() {
-        if var.fallback.is_none() {
-            let prop_id = OwnedPropId::from(var.name.ident.0.to_string());
-            if let Some(next) = var_defs.get(&prop_id).and_then(|defs| defs.last().copied()) {
-                return resolve_to_raw_value(next, var_defs, depth + 1);
-            }
-        }
-    }
-    Some(prop.value.raw.as_str().to_string())
-}
-
-fn swatch_markdown(color: &lsp_types::Color) -> String {
-    let r = (color.red * 255.0).round() as u8;
-    let g = (color.green * 255.0).round() as u8;
-    let b = (color.blue * 255.0).round() as u8;
-    let a = color.alpha;
-    let svg = format!(
-        "<svg xmlns='http://www.w3.org/2000/svg' width='14' height='14'>\
-         <rect width='14' height='14' fill='rgba({r},{g},{b},{a})' \
-         stroke='rgba(0,0,0,0.2)' stroke-width='0.5'/></svg>"
-    );
-    let encoded = BASE64.encode(svg.as_bytes());
-    format!("![](data:image/svg+xml;base64,{encoded})")
 }
 
 #[cfg(test)]
@@ -339,19 +270,6 @@ mod tests {
             text.contains("`#00ff00`"),
             "expected resolved value: {text}"
         );
-    }
-
-    #[test]
-    fn separator_picks_hard_break_for_helix() {
-        assert_eq!(multi_def_separator(Some("helix")), "  \n");
-        assert_eq!(multi_def_separator(Some("Helix")), "  \n");
-    }
-
-    #[test]
-    fn separator_defaults_to_paragraph_break() {
-        assert_eq!(multi_def_separator(None), "\n\n");
-        assert_eq!(multi_def_separator(Some("Visual Studio Code")), "\n\n");
-        assert_eq!(multi_def_separator(Some("Zed")), "\n\n");
     }
 
     #[test]
